Add IP to console banner (#13)
Add an /etc/issue with the default text from Arch, plus some pikvm
explanatory text, plus the IP address of the first fully configured
network interface.

This is useful for being able to immediately see the IP that the device
has been assigned to more quickly get to the web interface as long as
someone is able to plug it into a monitor. This is more convenient than
logging into the device because the device might be in OTG mode and
unable to accept keyboard input.