Remove data=journal from /etc/fstab
Data journaling mode from ext4 writes the data twice - in the journal and
non-journal area - for enhanced integrity.

While this may be acceptable for PST (Persistent storage), this is an
overkill for MSD (Mass Storage Drive) and leads to much slower writes.

Even for PST, this is an overkill for the following reasons:
 - The regular root drive is not using data journal mode, where integrity
   matters the most.
 - The dirty writeback policies of pikvm makes it extremely unlikely for a
   data corruption to occur.

Switch to the ext4's default journaling mode, data=ordered, which journals
metadata.

Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>